GH-750 节点无法安装
* 修复节点无法安装

* 格式化代码

* 删除多余配置文件

* 将yarn的超时时间延长

* 格式化代码

+++ spark-yun-backend/spark-yun-main/src/main/resources/bash/agent-standalone.sh
@@ -1,23 +1,19 @@
 #!/bin/bash
 
-######################
-# 检测安装环境脚本
-######################
+##########################################
+# 安装前,检测standalone模式是否允许安装
+##########################################
 
+# 获取脚本文件当前路径
+BASE_PATH=$(cd "$(dirname "$0")" || exit ; pwd)
+
+# 初始化环境变量
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
     source /etc/profile
     source ~/.bashrc
 elif [[ "$OSTYPE" == "darwin"* ]]; then
     source /etc/profile
     source ~/.zshrc
-elif [[ "$OSTYPE" == "cygwin" || "$OSTYPE" == "msys" || "$OSTYPE" == "mingw64" ]]; then
-    json_output="{ \
-                          \"status\": \"INSTALL_ERROR\", \
-                          \"log\": \"windows系统不支持安装\" \
-                        }"
-          echo $json_output
-          rm ${BASE_PATH}/agent-standalone.sh
-          exit 0
 else
     json_output="{ \
                       \"status\": \"INSTALL_ERROR\", \
@@ -28,8 +24,7 @@ else
       exit 0
 fi
 
-BASE_PATH=$(cd "$(dirname "$0")" || exit ; pwd)
-
+# 获取外部参数
 home_path=""
 agent_port=""
 spark_local="false"
@@ -42,14 +37,29 @@ for arg in "$@"; do
   esac
 done
 
+# 初始化agent_path
+agent_path="${home_path}/zhiqingyun-agent"
+
 # 判断home_path目录是否存在
-if [ ! -d "$home_path" ]; then
-  mkdir -p $home_path
+if [ ! -d "${agent_path}" ]; then
+  mkdir -p "${agent_path}"
 fi
 
+# 帮助用户初始化agent-env.sh文件
+if [ ! -f "${agent_path}/conf/agent-env.sh" ]; then
+  mkdir "${agent_path}/conf"
+  touch "${agent_path}/conf/agent-env.sh"
+  echo '#export JAVA_HOME=' >> "${agent_path}/conf/agent-env.sh"
+  echo '#export HADOOP_HOME=' >> "${agent_path}/conf/agent-env.sh"
+  echo '#export HADOOP_CONF_DIR=' >> "${agent_path}/conf/agent-env.sh"
+fi
+
+# 导入用户自己配置的环境变量
+source "${agent_path}/conf/agent-env.sh"
+
 # 判断是否之前已安装代理
-if [ -e "${home_path}/zhiqingyun-agent.pid" ]; then
-  pid=$(cat "${home_path}/zhiqingyun-agent.pid")
+if [ -e "${agent_path}/zhiqingyun-agent.pid" ]; then
+  pid=$(cat "${agent_path}/zhiqingyun-agent.pid")
   if ps -p $pid >/dev/null 2>&1; then
     json_output="{ \
             \"status\": \"RUNNING\", \
@@ -82,39 +92,19 @@ fi
 
 # 判断是否有java命令
 if ! command -v java &>/dev/null; then
-  json_output="{ \
+  # 如果没有java命令,判断用户是否配置了JAVA_HOME环境变量
+  if [ ! -n "$JAVA_HOME" ]; then
+    json_output="{ \
     \"status\": \"INSTALL_ERROR\", \
-    \"log\": \"未检测到java命令，需要java1.8.x环境,推荐命令: sudo yum install java-1.8.0-openjdk-devel java-1.8.0-openjdk -y \" \
-  }"
-  echo $json_output
-  rm ${BASE_PATH}/agent-standalone.sh
-  exit 0
+    \"log\": \"未检测到java1.8.x环境,节点请安装java 推荐命令: sudo yum install java-1.8.0-openjdk-devel java-1.8.0-openjdk -y,或者配置 ${agent_path}/conf/agent-env.sh文件中的JAVA_HOME变量\" \
+    }"
+    echo $json_output
+    rm ${BASE_PATH}/agent-standalone.sh
+    exit 0
+  fi
 fi
 
-# 判断java版本是否为1.8
-#java_version=$(java -version 2>&1 | awk -F '"' '/version/ {print $2}')
-#if [[ "$java_version" != "1.8"* ]]; then
-#  json_output="{ \
-#      \"status\": \"INSTALL_ERROR\", \
-#      \"log\": \"未检测到java1.8.x环境\" \
-#    }"
-#  echo $json_output
-#  rm ${BASE_PATH}/agent-standalone.sh
-#  exit 0
-#fi
-
-# 判断mpstat命令
-#if ! command -v mpstat &>/dev/null; then
-#  json_output="{ \
-#        \"status\": \"INSTALL_ERROR\", \
-#        \"log\": \"未检测到mpstat命令\" \
-#      }"
-#  echo $json_output
-#  rm /tmp/sy-env-kubernetes.sh
-#  exit 0
-#fi
-
-# 判断端口号是否被占用
+# 判断代理端口号是否被占用
 if ! netstat -tln | awk '$4 ~ /:'"$agent_port"'$/ {exit 1}'; then
   json_output="{ \
           \"status\": \"INSTALL_ERROR\", \
@@ -125,11 +115,11 @@ if ! netstat -tln | awk '$4 ~ /:'"$agent_port"'$/ {exit 1}'; then
   exit 0
 fi
 
+# 判断是否要默认安装spark
 if [ ${spark_local} = "false" ]; then
-  # 判断是否安装spark，判断是否配置SPARK_HOME
-  if [ -n "$SPARK_HOME" ]; then
-    SPARK_PATH=$SPARK_HOME
-  else
+
+  # 如果用户指定spark,则必须配置SPARK_HOME
+  if [ -z "$SPARK_HOME" ]; then
     json_output="{ \
               \"status\": \"INSTALL_ERROR\", \
               \"log\": \"未配置SPARK_HOME环境变量\" \
@@ -139,12 +129,11 @@ if [ ${spark_local} = "false" ]; then
     exit 0
   fi
 
-  # 从$SPARK_HOME/conf/spark-defaults.conf 配置文件中，获取SPARK_MASTER_PORT和SPARK_MASTER_WEBUI_PORT
-  # spark.master          spark://isxcode:7077
-  # spark.master.web.url  http://isxcode:8081
-  config_file="$SPARK_PATH/conf/spark-defaults.conf"
+  # 获取spark的配置文件
+  spark_config_path="$SPARK_HOME/conf/spark-defaults.conf"
 
-  if [ ! -f "$config_file" ]; then
+  # 检查spark配置文件是否存在
+  if [ ! -f "${spark_config_path}" ]; then
       json_output="{ \
                          \"status\": \"INSTALL_ERROR\", \
                          \"log\": \"$SPARK_HOME/conf/spark-defaults.conf配置文件不存在\" \
@@ -154,9 +143,8 @@ if [ ${spark_local} = "false" ]; then
         exit 0
   fi
 
-  spark_master_web_url=$(grep -E "^spark.master\.web\.url[[:space:]]+" "$config_file" | awk '{print $2}')
-  spark_master=$(grep -E "^spark.master[[:space:]]+" "$config_file" | awk '{print $2}')
-
+  # 检查配置文件中是否配置了spark.master.web.url
+  spark_master_web_url=$(grep -E "^spark.master\.web\.url[[:space:]]+" "$spark_config_path" | awk '{print $2}')
   if [[ -z $spark_master_web_url ]]; then
     json_output="{ \
                  \"status\": \"INSTALL_ERROR\", \
@@ -167,6 +155,8 @@ if [ ${spark_local} = "false" ]; then
     exit 0
   fi
 
+  # 检查配置文件中是否配置了spark.master
+  spark_master=$(grep -E "^spark.master[[:space:]]+" "$spark_config_path" | awk '{print $2}')
   if [[ -z $spark_master ]]; then
     json_output="{ \
                      \"status\": \"INSTALL_ERROR\", \
@@ -188,7 +178,7 @@ if [ ${spark_local} = "false" ]; then
                         \"log\": \"无法访问spark.master.web.url的服务，请检查spark服务。\" \
                       }"
       echo $json_output
-      rm /tmp/sy-env-standalone.sh
+      rm ${BASE_PATH}/agent-standalone.sh
       exit 0
     fi
   else
